folder Button renamed for ShowAllButton
