+++ src/features/movies/MoviePage/index.js
@@ -11,9 +11,9 @@ import { Section } from '../../../common/components/Section';
 import { Title } from '../../../common/components/Title';
 import { ErrorMessage } from '../../../common/components/ErrorMessage';
 import { PersonTile } from '../../../common/components/Tiles/PersonTile';
-import { StyledButton, Wrapper } from "../../../common/components/Button/styled";
-import { ReactComponent as ArrowDown } from "../../../common/components/Button/Arrow_down.svg";
-import { ReactComponent as ArrowUp } from "../../../common/components/Button/Arrow_up.svg";
+import { StyledButton, Wrapper } from "../../../common/components/ShowAllButton/styled";
+import { ReactComponent as ArrowDown } from "../../../common/components/ShowAllButton/Arrow_down.svg";
+import { ReactComponent as ArrowUp } from "../../../common/components/ShowAllButton/Arrow_up.svg";
 
 const MoviePage = () => {
     const [isShownAll, setIsShownAll] = useState(false);
